+++ src/go/configgenerator/filter_gen_path_rewrite.go
@@ -1,0 +1,63 @@
+// Copyright 2021 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package configgenerator
+
+import (
+	"fmt"
+
+	ci "github.com/GoogleCloudPlatform/esp-v2/src/go/configinfo"
+	"github.com/GoogleCloudPlatform/esp-v2/src/go/util"
+	"github.com/GoogleCloudPlatform/esp-v2/src/go/util/httppattern"
+	hcmpb "github.com/envoyproxy/go-control-plane/envoy/extensions/filters/network/http_connection_manager/v3"
+	"github.com/golang/protobuf/ptypes"
+	anypb "github.com/golang/protobuf/ptypes/any"
+)
+
+var prPerRouteFilterConfigGen = func(method *ci.MethodInfo, httpRule *httppattern.Pattern) (*anypb.Any, error) {
+	pr := MakePathRewriteConfig(method, httpRule)
+	if pr == nil {
+		return nil, nil
+	}
+
+	prAny, err := ptypes.MarshalAny(pr)
+	if err != nil {
+		return nil, fmt.Errorf("error marshaling path_rewrite per-route config to Any: %v", err)
+	}
+	return prAny, nil
+}
+
+var prFilterGenFunc = func(sc *ci.ServiceInfo) (*hcmpb.HttpFilter, []*ci.MethodInfo, error) {
+	perRouteConfigRequiredMethods, needed := needPathRewrite(sc)
+	if !needed {
+		return nil, nil, nil
+	}
+	return &hcmpb.HttpFilter{
+		Name: util.PathRewrite,
+	}, perRouteConfigRequiredMethods, nil
+}
+
+func needPathRewrite(serviceInfo *ci.ServiceInfo) ([]*ci.MethodInfo, bool) {
+	needed := false
+	var perRouteConfigRequiredMethods []*ci.MethodInfo
+	for _, method := range serviceInfo.Methods {
+		for _, httpRule := range method.HttpRule {
+			if pr := MakePathRewriteConfig(method, httpRule); pr != nil {
+				needed = true
+				perRouteConfigRequiredMethods = append(perRouteConfigRequiredMethods, method)
+			}
+		}
+	}
+	return perRouteConfigRequiredMethods, needed
+}
